requirements.txt
alembic==1.5.2
aniso8601==8.1.0
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.19.5
psycopg2-binary==2.8.6
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
